INTYGFV-16352: Refactor

+++ certificate-service/app/src/main/java/se/inera/intyg/certificateservice/infrastructure/certificate/persistence/entity/CertificateEntity.java
@@ -36,7 +36,7 @@ public class CertificateEntity {
   @Column(name = "modified")
   private LocalDateTime modified;
   @Column(name = "version")
-  private int version;
+  private Long version;
   @OneToOne(fetch = FetchType.LAZY, cascade = {CascadeType.MERGE, CascadeType.PERSIST})
   private CertificateModelEntity certificateModel;
   @ManyToOne(fetch = FetchType.LAZY, cascade = {CascadeType.MERGE, CascadeType.PERSIST})
